Editar usuario

+++ app/Http/Controllers/UsuariosController.php
@@ -66,14 +66,15 @@ class UsuariosController extends Controller
         $usuario->nombre = $request->input("nombre");
         $usuario->password = Hash::make($request->input("password"));
         $usuario->estado = "0";
-        $usuario->correo=$request->input("correo");
-        $imagen= $request->file("imagen");
-        $ruta=Storage::disk('public_uploads')->put('',$imagen);
-        $usuario->imagenPerfil=$ruta;
+        $usuario->correo = $request->input("correo");
+        $imagen = $request->file("imagen");
+        $ruta = Storage::disk('public_uploads')->put('', $imagen);
+        $usuario->imagenPerfil = $ruta;
         $usuario->save();
         $token = JWTAuth::fromUser($usuario);
         return response()->json(compact('usuario', 'token'), 201);
     }
+
     //Agregar vendedor
     public function nuevoVendedor(Request $request)
     {
@@ -97,9 +98,9 @@ class UsuariosController extends Controller
         $usuario->estado = "0";
         $usuario->rolId = "1";
         $idvend = $usuario->id;
-        $imagen= $request->file("imagen");
-        $ruta=Storage::disk('public')->put('',$imagen);
-        $usuario->imagenPerfil=$ruta;
+        $imagen = $request->file("imagen");
+        $ruta = Storage::disk('public')->put('', $imagen);
+        $usuario->imagenPerfil = $ruta;
         $usuario->save();
         $token = JWTAuth::fromUser($usuario);
         //Se agrega al vendedor a la tabla vendedores con su vendedor ID
@@ -110,47 +111,53 @@ class UsuariosController extends Controller
         $vendedor->save();
         return response()->json(compact('usuario', 'token'), 201);
     }
+
     //VerUsuario
-    public function verUsuario($id){
-        $usuario=Usuario::find($id);
+    public function verUsuario($id)
+    {
+        $usuario = Usuario::find($id);
         return json_encode($usuario);
     }
-    public function edtarUsuario($id,Request $request){
+
+    public function edtarUsuario($id, Request $request)
+    {
         $validator = Validator::make($request->all(), [
             "nombre" => "required",
-            "password" => "required|min:6",
-            "rpassword" => "required|same:password"
+            "password" => "required|min:6"
         ], [
             "nombre.required" => "Por favor llenar el campo de nombre.",
             "password.required" => "Debe llenar el campo del password.",
-            "password.min" => "La contraseña debe ser de al menos 6 carácteres.",
-            "rpassword.same" => "La contraseña y su confirmación deben ser iguales."
+            "password.min" => "La contraseña debe ser de al menos 6 carácteres."
         ]);
 
         if ($validator->fails())
             return response()->json(["estado" => false, "detalle" => $validator->errors()->all()]);
-        $usuario=Usuario::find($id);
-        $imagenVieja=$usuario->imagenPerfil;
-        if(Storage::disk('public_uploads')->exists($imagenVieja)==null){
-            Storage::disk('public_uploads')->delete($imagenVieja);
-            $imagenNueva=$request->file('imagen');
-            $extension=$imagenNueva->getClientOriginalExtension();
-            $imagenRezize=Image::make($imagenNueva->getRealPath());
-            $imagenRezize->resize(150,150);
-            $imagenRezize->save(public_path("storage/".$id.".".$extension));
-            $usuario->imagenPerfil=$id.".".$extension;
-            $usuario->password = Hash::make($request->input("password"));
+        $usuario = Usuario::find($id);
+        $pass = $usuario->password;
+        if (Hash::check($request->input('password'), $pass)) {
+            $imagenVieja = $usuario->imagenPerfil;
+            if (Storage::disk('public_uploads')->exists($imagenVieja) == null) {
+                Storage::disk('public_uploads')->delete($imagenVieja);
+                $imagenNueva = $request->file('imagen');
+                $extension = $imagenNueva->getClientOriginalExtension();
+                $imagenRezize = Image::make($imagenNueva->getRealPath());
+                $imagenRezize->resize(150, 150);
+                $imagenRezize->save(public_path("storage/" . $id . "." . $extension));
+                $usuario->imagenPerfil = $id . "." . $extension;
+                $usuario->nombre=$request->input('nombre');
+                $usuario->save();
+                return json_encode(["estado" => true]);
+            }
+            $imagenNueva = $request->file('imagen');
+            $extension = $imagenNueva->getClientOriginalExtension();
+            $imagenRezize = Image::make($imagenNueva->getRealPath());
+            $imagenRezize->resize(150, 150);
+            $imagenRezize->save(public_path("storage/" . $id . "." . $extension));
+            $usuario->imagenPerfil = $id . "." . $extension;
+            $usuario->nombre=$request->input('nombre');
             $usuario->save();
-            return json_encode(["estado"=>true]);
+            return json_encode(["estado" => true]);
         }
-        $imagenNueva=$request->file('imagen');
-        $extension=$imagenNueva->getClientOriginalExtension();
-        $imagenRezize=Image::make($imagenNueva->getRealPath());
-        $imagenRezize->resize(150,150);
-        $imagenRezize->save(public_path("storage/".$id.".".$extension));
-        $usuario->imagenPerfil=$id.".".$extension;
-        $usuario->password = Hash::make($request->input("password"));
-        $usuario->save();
-        return json_encode(["estado"=>true]);
+        return json_encode(["estado" => false, "detalle" => ["Contraseña incorrecta"]]);
     }
 }
